Test enable interrupts
This loops 10,000 times outputting characters to the VGA display
and then continues running the startup code as before. If interrupts
are not enabled all works as planned, you see characters
fly by on the screen and then "OS returned!" is displayed in red in
the upper left hand corner.

If you edit boot.asm and enable interrupts by removing the comment on
"sti" instruction at line 30. Then a some characters will be output and
then it "reboots" back into BIOS and grub restarts. The same behavior
is seen both with "make run" on qemu-system-x86_64 and on my test
hardware after flashing with:

  sudo dd bs=4M of=/dev/sdb if=build/os-x86_64.iso ; sync